requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
scipy>=1.7.0
mne>=1.0.0
mlflow>=2.0.0
PyYAML>=6.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0
torch>=1.9.0
